Changed all integers to ints
All integers in function parameters and return types are now either int
or uint.  In fact most of the numeric values in the code are now either
int or uint, with the exception being coordinate values, which are
ubytes.  This has been done to reduce the amount of conversion necessary
to expand the codebase.

+++ src/fov.d
@@ -216,7 +216,7 @@ void TCOD_map_compute_fov_recursive_shadowcasting(
  +   viewer_y   = The y coordinate of the start point for field-of-vision
  +                calculation
  +/
-void calc_visible( Map* to_display, ushort viewer_x, ushort viewer_y )
+void calc_visible( Map* to_display, uint viewer_x, uint viewer_y )
 {
   TCOD_map_compute_fov_recursive_shadowcasting(
     to_display, viewer_x, viewer_y, 0, true );
